Remove EDH charities api methods & tests
