+++ source/api/charities/readme.md
@@ -26,5 +26,5 @@ A pending promise that will either resolve to:
 ```javascript
 import { searchCharities } from 'supporticon/api/charities'
 
-searchCharities({ campaign: 'au-0' })
+searchCharities({ campaign: 'f440df6c-1101-4331-ac78-4fc5bc276f4e' })
 ```
